cpp09: ex02_alt: major simplification of insertion function
Can now rely on the insertion block size being bounded for the worst
case.

The search range is now delimited by the insertion_block size or the end
of the container, whichever is smallest.

This allows for much simpler calculations and major reduction of local
variables.

This does however come at the expense of reducing the best case
comparisons.